semantic-kernel-sample/requirements.txt
semantic-kernel
python-dotenv
azure-ai-documentintelligence==1.0.2
openai
azure-core
azure-identity
streamlit

